requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.29.0
Babel==2.13.1
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-mail==1.4.1
Faker==20.1.0
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openai==0.28
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pip==23.3.1
psycopg-binary==3.1.13
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pyfcm==1.5.4
PyMySQL==1.1.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==69.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing-extensions==4.8.0
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0